Adding new command agent:add-identity

+++ src/AwsInspector/Ssh/Identity.php
@@ -7,9 +7,12 @@ class Identity
 
     protected $privateKey;
 
-    public function __construct(PrivateKey $privateKey)
+    protected $keepIdentity;
+
+    public function __construct(PrivateKey $privateKey, $keepIdentity=false)
     {
         $this->privateKey = $privateKey;
+        $this->keepIdentity = $keepIdentity;
     }
 
     public function getPrivateKeyFile() {
@@ -35,10 +38,12 @@ class Identity
 
     public function __destruct()
     {
-        $this->removeIdentity();
-        // Remove control paths for ssh multiplexing. See \AwsInspector\Ssh->__toString()
-        // TODO: deleting these files isn't enough. The mux needs to be closed.
-        exec('rm ~/mux* 2> /dev/null');
+        if (!$this->keepIdentity) {
+            $this->removeIdentity();
+            // Remove control paths for ssh multiplexing. See \AwsInspector\Ssh->__toString()
+            // TODO: deleting these files isn't enough. The mux needs to be closed.
+            exec('rm ~/mux* 2> /dev/null');
+        }
     }
 
 }
